server 객체 분리
